kvserver: put multiSSTWriter into sep file
Epic: CRDB-46488
Release note: none
