+++ pkg/kv/kvserver/BUILD.bazel
@@ -23,6 +23,7 @@ go_library(
         "merge_queue.go",
         "metric_rules.go",
         "metrics.go",
+        "multi_sst_writer.go",
         "mvcc_gc_queue.go",
         "queue.go",
         "queue_helpers_testutil.go",
